# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Forum Diskusi API Documentation",
        "description": "API documentation for the Final Project Forum Diskusi",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/moderation/reports": {
            "get": {
                "tags": [
                    "8. MODERASI"
                ],
                "summary": "List all reports (Moderator Only)",
                "operationId": "eee69615c80db6102036402b603facad",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "default": "pending",
                            "enum": [
                                "pending",
                                "resolved",
                                "dismissed"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of reports",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/moderation/reports/{id}/resolve": {
            "post": {
                "tags": [
                    "8. MODERASI"
                ],
                "summary": "Resolve a report (Moderator Only)",
                "operationId": "600f32efbbf4ed1f6dbfe244edeb6e49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Report ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "resolved",
                                            "dismissed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Report resolved successfully"
                    },
                    "404": {
                        "description": "Report not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/moderation/ban": {
            "post": {
                "tags": [
                    "8. MODERASI"
                ],
                "summary": "Ban a user (Moderator Only)",
                "operationId": "b4166dbf0c5c05e45227b35ec9ed7405",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "reason"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Violating community guidelines"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Banned for 30 days"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User banned successfully"
                    },
                    "403": {
                        "description": "Cannot ban an admin"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/moderation/warn": {
            "post": {
                "tags": [
                    "8. MODERASI"
                ],
                "summary": "Warn a user (Moderator Only)",
                "operationId": "b7ffe5c1cb147d02740caa2d7760b19e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "reason"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Inappropriate language"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User warned successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/moderation/unban": {
            "post": {
                "tags": [
                    "8. MODERASI"
                ],
                "summary": "Unban a user (Moderator Only)",
                "operationId": "3cafd052969f9e82cf563ef13325f71e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "reason"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Ban period ended or appeal accepted"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User unbanned successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/moderation/logs": {
            "get": {
                "tags": [
                    "8. MODERASI"
                ],
                "summary": "List all moderation logs (Moderator Only)",
                "operationId": "161de3dd590e0922364a0d778fdeba74",
                "responses": {
                    "200": {
                        "description": "List of logs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}